wesdash_refresh_v2/requirements.txt
requests>=2.31
pandas>=2.1
pyyaml>=6.0
python-dotenv>=1.0
openpyxl>=3.1
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
fiona>=1.9
pyarrow>=14.0
